OpenCL: add support for OpenCL (HPC)
It seems to me that it is possible to use GPU acceleration via OpenCL
within flatpak:
- https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1535
- https://github.com/flathub/org.freedesktop.Platform.GL.nvidia

However, it's not clear what is actually needed for this to work (poor
documentation? or I'm failing to find it?).
- what is the minimum required to run `cl-info` inside a flatpak?
  - john is a CLI (HPC) application.

I don't need to worry too much about AMD/Intel/everybody else for now.
NVIDIA is the target at the moment.

Our community does not have suitable hardware with CUDA installed to
use/test this PR.

So let's tag this commit as a WIP, for experimentation.

=> Review/Revert/Improve:
```
+        "--socket=x11",
+        "--device=all"
```

. Should I also add org.freedesktop.Platform.HIP.Nvidia?
. Is the `add-extensions:` necessary or useful? I want to enable NVIDIA support
but allow poor people (like me) to continue using flatpak, e.g., for CPU
formats.

Signed-off-by: Claudio André <dev@claudioandre.slmail.me>